added star/ctx/ideal and un/paired modes to featureCounts

+++ scripts/process_featureCounts.sh
@@ -15,7 +15,7 @@ if [[ ${PIPESTATUS[0]} -ne 4 ]]; then
 fi
 
 OPTIONS=
-LONGOPTS=index:,gtf:,pdata:,out:,nthread:,log:,dexseq
+LONGOPTS=index:,gtf:,pdata:,out:,nthread:,log:,dexseq,hisat2,star,contextmap,ideal,paired,unpaired
 
 # -regarding ! and PIPESTATUS see above
 # -temporarily store output to be able to check for errors
@@ -30,7 +30,16 @@ fi
 # read getopt’s output this way to handle the quoting right:
 eval set -- "$PARSED"
 
+declare -A map
+declare -A paths
+
+map[hisat]=n paths[hisat]="HISAT/dta"
+map[star]=n paths[star]="STAR/quant"
+map[contextmap]=n paths[contextmap]="CONTEXTMAP"
+map[ideal]=n paths[ideal]="IDEAL"
 dexseq=n
+paired=''
+
 # now enjoy the options in order and nicely split until we see --
 while true; do
     case "$1" in
@@ -62,6 +71,26 @@ while true; do
 			dexseq="y"
 			shift
 			;;
+		--hisat2)
+            map[hisat]=y
+            shift
+            ;;
+        --star)
+            map[star]=y
+            shift
+            ;;
+        --contextmap)
+            map[contextmap]=y
+            shift
+            ;;
+        --ideal)
+            map[ideal]=y
+            shift
+            ;;
+		--paired)
+			paired="-p"
+			shift
+			;;
         --)
             shift
             break
@@ -72,41 +101,47 @@ while true; do
     esac
 done
 
-basein=$out/HISAT/dta
-baseout=$out/COUNTS/featureCounts
 featureCounts='/home/software/subread/bin/featureCounts'
 
 dir=$(basename $out)
-
 mkdir -p $out/COUNTS
 
-[ -f "$baseout" ] && echo "[INFO] [featureCounts] $baseout already exists; skipping.."$'\n' && exit 
-
-hisatBams=()
+bams=()
 for sample in `sed '1d' $pdata | cut -f1`; do
-	hisatBams+=( "$sample.bam" )
+	bams+=( "$sample.bam" )
 done
 
-echo "[INFO] [featureCounts] ["`date "+%Y/%m/%d-%H:%M:%S"`"] Started processing $baseout"$'\n'
-##gene level counting
-cd $out/HISAT/dta
-watch pidstat -dru -hlH '>>' $log/featureCounts_${dir}-$(date +%s).pidstat & wid=$!
 
-$featureCounts -T $nthread -p -B -C -a $gtf --primary -o $baseout ${hisatBams[@]}
+for method in "hisat" "star" "contextmap" "ideal"; do
 
-kill -15 $wid
+	if [[ "${map[$method]}}" = "y" ]]; then
+		basein=$out/${paths[$method]}
+		baseout=$out/COUNTS/featureCounts.$method
 
-##exon_part level counting
-if [[ "$dexseq" = "y" ]]; then
-	echo "[INFO] [featureCounts] ["`date "+%Y/%m/%d-%H:%M:%S"`"] Started processing $baseout.DEXSeq"$'\n'
-	watch pidstat -dru -hlH '>>' $log/featureCounts_exonicpart_${dir}-$(date +%s).pidstat & wid=$!
+		[ -f "$baseout" ] && echo "[INFO] [featureCounts] $baseout already exists; skipping.."$'\n' && continue
+		cd $basein
+
+		##gene level counting
+		echo "[INFO] [featureCounts] ["`date "+%Y/%m/%d-%H:%M:%S"`"] Started processing $baseout"$'\n'
+		watch pidstat -dru -hlH '>>' $log/featureCounts_${dir}-$(date +%s).pidstat & wid=$!
+
+		$featureCounts -T $nthread $p -B -C -a $gtf --primary -o $baseout ${bams[@]}
+
+		kill -15 $wid
+		sed '1d' $baseout | cut -f 1,7- > $out/gene.counts.$method
+
+		##exon_part level counting
+		if [[ "$dexseq" = "y" ]]; then
+			echo "[INFO] [featureCounts] ["`date "+%Y/%m/%d-%H:%M:%S"`"] Started processing $baseout.DEXSeq"$'\n'
+			watch pidstat -dru -hlH '>>' $log/featureCounts_exonicpart_${dir}-$(date +%s).pidstat & wid=$!
 	
-	$featureCounts -T $nthread -p -B -C -O -f -t exonic_part -a $index/dexseq/annot.noaggregate.gtf --primary -o $baseout.DEXSeq ${hisatBams[@]}
+			$featureCounts -T $nthread $p -B -C -O -f -t exonic_part -a $index/dexseq/annot.noaggregate.gtf --primary -o $baseout.DEXSeq ${bams[@]}
 
-	kill -15 $wid
-fi
+			kill -15 $wid
+		fi
+	fi
+done
 
-sed '1d' $baseout | cut -f 1,7- > $out/gene.counts
 
 echo "[INFO] [featureCounts] ["`date "+%Y/%m/%d-%H:%M:%S"`"] Finished processing $out"$'\n'
